+++ src/main/java/com/dionialves/snapdogdelivery/domain/admin/product/dto/ProductResponseDTO.java
@@ -3,6 +3,7 @@ package com.dionialves.snapdogdelivery.domain.admin.product.dto;
 import java.math.BigDecimal;
 
 import com.dionialves.snapdogdelivery.domain.admin.product.Product;
+import com.dionialves.snapdogdelivery.domain.admin.product.ProductCategory;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -21,16 +22,21 @@ public class ProductResponseDTO {
     private String description;
     private String imageUrl;
     private boolean active;
+    private ProductCategory category;
+    private String categoryLabel;
 
     public static ProductResponseDTO fromEntity(Product product) {
 
+        var category = product.getCategory();
         return new ProductResponseDTO(
                 product.getId(),
                 product.getName(),
                 product.getPrice(),
                 product.getDescription(),
                 product.getImageUrl(),
-                product.isActive());
+                product.isActive(),
+                category,
+                category != null ? category.getLabel() : null);
 
     }
 }